# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "CodeRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "CodeRequest",
        "type": "object"
      },
...
      "Command": {
        "properties": {
          "at": {
            "description": "Seconds since local midnight of the run day.",
            "minimum": 0.0,
            "title": "At",
            "type": "number"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "type": {
            "description": "break_asset | dispatch",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "at",
          "type"
        ],
        "title": "Command",
        "type": "object"
      },
...
      "DaysRequest": {
        "description": "A range of run days, each replayed without commands (the viewer's +1 week / +1 month).",
        "properties": {
          "from": {
            "description": "First run day (YYYY-MM-DD, local).",
            "title": "From",
            "type": "string"
          },
          "m2c": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "The meter-to-cash run; only its `seed` matters here (the days' background incidents never depend on its field orders).",
            "title": "M2C"
          },
          "seed": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Run seed, as for /api/sim/timeline.",
            "title": "Seed"
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Overrides for the run settings, as for /api/sim/timeline.",
            "title": "Settings"
          },
          "to": {
            "description": "Last run day, inclusive; at most 62 days from the first.",
            "title": "To",
            "type": "string"
          },
          "town": {
            "description": "Pack preset (e.g. 'small_town') or town id.",
            "title": "Town",
            "type": "string"
          }
        },
        "required": [
          "town",
          "from",
          "to"
        ],
        "title": "DaysRequest",
        "type": "object"
      },
...
      "FrameRequest": {
        "properties": {
          "at": {
            "description": "Seconds since local midnight of the run day.",
            "minimum": 0.0,
            "title": "At",
            "type": "number"
          },
          "commands": {
            "items": {
              "$ref": "#/components/schemas/Command"
            },
            "maxItems": 500,
            "title": "Commands",
            "type": "array"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Run day (local); default: the town's scenario date.",
            "title": "Date"
          },
          "m2c": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "The meter-to-cash run ({settings, actions, outages, seed}) whose field orders for this day become crew jobs (see /api/m2c/*).",
            "title": "M2C"
          },
          "premises": {
            "default": true,
            "title": "Premises",
            "type": "boolean"
          },
          "seed": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Run seed: re-rolls the day's background incidents (default: the m2c run's seed, else none: the town's draws).",
            "title": "Seed"
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Overrides for the run settings: timings, crews, incident rates (see GET /api/sim/settings/schema?town=).",
            "title": "Settings"
          },
          "town": {
            "description": "Pack preset (e.g. 'small_town') or town id.",
            "title": "Town",
            "type": "string"
          }
        },
        "required": [
          "town",
          "at"
        ],
        "title": "FrameRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TimelineRequest": {
        "properties": {
          "commands": {
            "items": {
              "$ref": "#/components/schemas/Command"
            },
            "maxItems": 500,
            "title": "Commands",
            "type": "array"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Run day (local); default: the town's scenario date.",
            "title": "Date"
          },
          "m2c": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "The meter-to-cash run ({settings, actions, outages, seed}) whose field orders for this day become crew jobs (see /api/m2c/*).",
            "title": "M2C"
          },
          "seed": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Run seed: re-rolls the day's background incidents (default: the m2c run's seed, else none: the town's draws).",
            "title": "Seed"
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Overrides for the run settings: timings, crews, incident rates (see GET /api/sim/settings/schema?town=).",
            "title": "Settings"
          },
          "town": {
            "description": "Pack preset (e.g. 'small_town') or town id.",
            "title": "Town",
            "type": "string"
          }
        },
        "required": [
          "town"
        ],
        "title": "TimelineRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Seeded utility-town engine: geography, networks, customers, simulation, meter-to-cash.",
    "title": "utilsim",
    "version": "0.10.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/share/decode": {
      "post": {
        "description": "The validated proposal a code holds, in the shape ``POST /api/setup-agent/validate`` answers with.",
        "operationId": "share_decode_api_share_decode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Share Decode"
      }
    },
    "/api/share/encode": {
      "post": {
        "description": "The code for a simulation's inputs: a wizard proposal by alias (``summary`` may be left out), with a Studio\nrecord's ``townRef`` and flat ``opsSettings`` accepted in place of ``townOverrides`` and ``operations``.",
        "operationId": "share_encode_api_share_encode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Proposal",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Share Encode"
      }
    },
    "/api/sim/days": {
      "post": {
        "description": "``utility-days/1.0``: the run days from ``from`` to ``to`` (inclusive, at most 62), each replayed with no\ncommands: per day its ``date``, ``interruptions`` (as the timeline reports them: the meter-to-cash run's\n``outages`` for the days the viewer skips over) and counts of ``incidents`` and ``jobs``. The days' background\nincidents are the same draws as a single-day timeline's, so the two always agree.",
        "operationId": "post_days_api_sim_days_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DaysRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Days"
      }
    },
    "/api/sim/frame": {
      "post": {
        "description": "One complete ``utility-state/1.0`` frame at ``at`` with the run's incidents applied (open switches, closed\nvalves, leaks); ``premises.unsupplied`` lists customers without supply.",
        "operationId": "post_frame_api_sim_frame_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FrameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Frame"
      }
    },
    "/api/sim/settings": {
      "get": {
        "description": "Default run settings for operations runs (override per request with ``settings``); ``?town=`` gives that\ntown's (its crews, shift, targets, voltage floor and incident rates).",
        "operationId": "get_settings_api_sim_settings_get",
        "parameters": [
          {
            "in": "query",
            "name": "town",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Town"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Settings"
      }
    },
    "/api/sim/settings/schema": {
      "get": {
        "description": "Operations settings as JSON Schema (titles, units, bounds, effects; ``x-town`` names the town config field a\ndefault comes from) with their defaults, for the viewer's Configuration. ``?town=`` fills the defaults from that\ntown. Send overrides as a timeline or frame request's ``settings``.",
        "operationId": "get_settings_schema_api_sim_settings_schema_get",
        "parameters": [
          {
            "in": "query",
            "name": "town",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Town"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Settings Schema"
      }
    },
    "/api/sim/timeline": {
      "post": {
        "description": "``utility-timeline/1.0``: incidents, crew jobs with road routes, events, state changes and field-visit reads\nfor a run's command list. Appending a command never changes what earlier commands produced.",
        "operationId": "post_timeline_api_sim_timeline_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Timeline"
      }
    },
...
  }
}